Ticket #47669 - Retro Changelog Plugin accepts invalid value in nsslapd-changelogmaxage attribute
Description: Converted parse_duration to the private api and applied the
following config attributes.
  cn=changelog5,cn=config
    nsslapd-changelogmaxage
    nsslapd-changelogtrim-interval
    nsslapd-changelogcompactdb-interval
  cn=Retro Changelog Plugin,cn=plugins,cn=config
    nsslapd-changelogmaxage
These config attributes take unsigned integer as well as digit[SsMmHhDdWw].
Other input are logged in the error log and the invalid value is ignored.

https://fedorahosted.org/389/ticket/47669

Reviewed by rmeggins@redhat.com (Thank you, Rich!)

+++ ldap/servers/slapd/value.c
@@ -532,7 +532,7 @@ slapi_value_get_timelong(const Slapi_Value *value)
         p = slapi_ch_malloc(value->bv.bv_len + 1);
         memcpy (p, value->bv.bv_val, value->bv.bv_len);
         p [value->bv.bv_len] = '\0';
-        r = (long)parse_duration(p);
+        r = parse_duration(p);
         slapi_ch_free((void **)&p);
 	}
 	return r;
